Create shared/components folder, Move FormField to shared/components/ Remove FormFieldInterface from SignupFormInterface
