+++ src/components/SignupForm/interfaces/SignupFormInterfaces.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 export interface SignUpFormDataInterface {
   name: string;
   email: string;
@@ -15,11 +13,6 @@ export interface FormFieldConfigInterface {
   errorText: string;
   value: string;
 }
-
-export type FormFieldInterface = {
-  onClick: (event: React.MouseEvent<HTMLDivElement>) => void;
-  onChange: (event: React.ChangeEvent<HTMLInputElement>) => void;
-} & FormFieldConfigInterface;
 
 export interface ErrorInterface {
   isError: boolean;
