+++ src/components/SignupForm/interfaces/ValidationInterface.tsx
@@ -1,10 +1,8 @@
 import React, { SetStateAction } from 'react';
 import { ErrorInterface } from './SignupFormInterfaces';
 
-interface ValidationInterface {
+export interface ValidationInterface {
   error: ErrorInterface;
   setError: React.Dispatch<SetStateAction<ErrorInterface>>;
   validateForm: () => boolean;
 }
-
-export default ValidationInterface;
